refactor: migrate financial calculations from double to decimal - Changed payroll and financial variables from double to decimal to prevent precision loss. - Updated BANGLUONG logic, DASHBOARD_DTO, and related report forms to use decimal type natively. - Added UI form (FrmOllamaConfig) to configure AI parameters (Temperature, Max Tokens, Context Window, Top K, Top P, Repeat Penalty). - Implemented validation for AI configuration inputs and linked them to OllamaService.

+++ QLyNSu/FORM_BAOCAO/FrmBaoCaoChiTiet.cs
@@ -166,11 +166,11 @@ namespace QLyNSu.FORM_BAOCAO
                 int thang = (int)makycong % 100;
 
                 // Query overtime hours
-                double otHours = 0;
+                decimal otHours = 0;
                 var lstTangCa = db.TB_TANGCA.Where(x => x.MANV == manv && x.THANG == thang && x.NAM == nam).ToList();
                 if (lstTangCa.Count > 0)
                 {
-                    otHours = (double)lstTangCa.Sum(x => x.SOGIO ?? 0);
+                    otHours = (decimal)lstTangCa.Sum(x => x.SOGIO ?? 0);
                 }
 
                 // Query allowances
